+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.0.4
 +   Import/export account private keys as a word phrase
++   Convex Wallet tool
 
 ### 0.0.3
 +   Pre-alpha release